Cache user list on every (dis)connect
When a user connects to chat, the server delivers them a JSON-encoded
string that contains the number of connections it's managing and an
array of all connected users. We now store the string in Redis as well
so we can figure out which users are in chat from destinygg/website.